OColecionadorClassifier/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
minio==7.1.0
tensorflow==2.6.0
numpy==1.19.5
pillow==8.4.0
